+++ src/render/DebugRenderer.cpp
@@ -19,7 +19,7 @@ using namespace wheels;
 namespace
 {
 
-enum BindingSet : uint32_t
+enum BindingSet : uint8_t
 {
     CameraBindingSet,
     GeometryBuffersBindingSet,
