+++ src/render/ForwardRenderer.cpp
@@ -28,7 +28,7 @@ using namespace wheels;
 namespace
 {
 
-enum BindingSet : uint32_t
+enum BindingSet : uint8_t
 {
     LightsBindingSet,
     LightClustersBindingSet,
