+++ src/render/SkyboxRenderer.cpp
@@ -18,7 +18,7 @@ using namespace wheels;
 namespace
 {
 
-enum BindingSet : uint32_t
+enum BindingSet : uint8_t
 {
     SkyboxBindingSet,
     CameraBindingSet,
